add dynamic container discovery
When merging metrics from containers scaled with docker-compose and
single exporter-merger, it is impossible to react to docker scaling
without reconfiguring exporter-merger.

To help with this, we introduce a concept of static and dynamic
discovery. Static discovery is left "as is", allowing scrape URLs to
be statically set. Dynamic discovery uses Docker SDK to query running
containers and match against a set of configured labels with port and
endpoint configuration specified.

The usage is highlighted in README, snippet:

```yaml
environment:
- MERGER_LABEL=traefik.frontend.rule
- MERGER_LABEL_CONFIG=/metrics:8080:Host:my-container /metrics:8080:Host:my-container2
```

This will cause all containers with `traefik.frontend.rule` label with
values `Host:my-container` and `Host:my-container2` to be scraped at
/metrics:8080.

Signed-off-by: Martin Polednik <m.polednik@gmail.com>